Automation request rest api
